requirements.txt
Django==5.0.6
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
openpyxl==3.1.5
pandas==2.2.2
beautifulsoup4==4.12.3